requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
boto3==1.26.122
botocore==1.29.122
build==1.2.2.post1
cachetools==5.5.1
celery==5.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
Deprecated==1.2.18
dnspython==2.7.0
durationpy==0.9
faiss-cpu==1.9.0.post1
fastapi==0.115.7
filelock==3.17.0
Flask==2.3.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
kombu==5.4.2
kubernetes==32.0.0
langchain==0.0.252
langfuse==1.0.18
langsmith==0.0.92
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.25.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pillow==11.1.0
pinecone==5.4.2
pinecone-client==2.2.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
posthog==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
pulsar-client==3.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
Pygments==2.19.1
pypdf==3.15.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.4
PyYAML==6.0.2
redis==5.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.6.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uuid==1.30
uvicorn==0.34.0
vine==5.1.0
watchdog==3.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.3.1
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
